# server/src/docs/swagger.yaml
definitions:
...
  models.File:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      id:
        type: integer
      object_key:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Guide:
    properties:
      author:
        type: string
      author_image_url:
        type: string
      created_at:
        type: string
      date:
        type: string
      full_text:
        type: string
      guide_name:
        type: string
      id:
        type: integer
      mins_read:
        type: integer
      sub_title:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - author
    - author_image_url
    - full_text
    - guide_name
    - sub_title
    - title
    type: object
  models.Persona:
    properties:
      created_at:
        type: string
      id:
        type: integer
      persona_description:
        type: string
      persona_title:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
    required:
    - persona_description
    - persona_title
    type: object
...
  models.Tag:
    properties:
      created_at:
        type: string
      files:
        items:
          $ref: '#/definitions/models.File'
        type: array
      id:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      id:
        type: integer
      personas:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      task_description:
        type: string
      task_name:
        type: string
      updated_at:
        type: string
    required:
    - task_description
    - task_name
    type: object
...
info:
  contact:
    email: oduneye.d@northeastern.edu and dupuguntla@northeastern.edu
    name: David Oduneye and Akshay Dupuguntla
  description: Backend Server for Legacy App
  title: Legacy API
  version: "1.0"
paths:
  /api/{fid}/filename:
    get:
      description: Returns a filename from file id
      operationId: get-filename
      parameters:
      - description: FileID
        in: path
        name: fid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "404":
          description: Failed to fetch file
          schema:
            type: string
      summary: Gets a filename from file id
      tags:
      - file
  /api/files/:
    get:
      description: Returns all file information
      operationId: get-all-files
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.File'
            type: array
        "404":
          description: Failed to fetch files
          schema:
            type: string
      summary: Gets all file information
      tags:
      - file
  /api/files/{fid}:
    delete:
      description: Deletes a file from file id
      operationId: delete-file
      parameters:
      - description: FileID
        in: path
        name: fid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File deleted
          schema:
            type: string
        "404":
          description: Failed to delete file
          schema:
            type: string
      summary: Deletes a file from file id
      tags:
      - file
    get:
      description: Returns a url to download a file from file id for a certain amount
        of days
      operationId: get-file
      parameters:
      - description: FileID
        in: path
        name: fid
        required: true
        type: string
      - description: Days
        in: query
        name: days
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "404":
          description: Failed to get presigned url
          schema:
            type: string
      summary: Gets a url to download a file from file id for a certain amount of
        days
      tags:
      - file
  /api/files/{uid}:
    post:
      consumes:
      - application/json
      description: Creates a file
      operationId: create-file
      parameters:
      - description: UserID
        in: path
        name: uid
        required: true
        type: string
      - description: File
        in: body
        name: file
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "400":
          description: Failed to create file
          schema:
            type: string
      summary: Creates a file
      tags:
      - file
  /api/files/{uid}/user:
    get:
      description: Returns all file information from user id
      operationId: get-all-user-files
      parameters:
      - description: UserID
        in: path
        name: uid
        required: true
        type: string
      - description: Tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.File'
            type: array
        "404":
          description: Failed to fetch files
          schema:
            type: string
      summary: Gets all file information from user id
      tags:
      - file
  /api/files/makepdf/{uid}:
    post:
      description: Returns a pdf from a json file
      operationId: generate-pdf
      parameters:
      - description: UserID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "404":
          description: Failed to generate pdf
          schema:
            type: string
      summary: Generates a pdf from a json file
      tags:
      - file
  /api/guides/:
    get:
      description: Returns all guides
      operationId: get-all-guides
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Guide'
            type: array
        "404":
          description: Failed to fetch guides
          schema:
            type: string
      summary: Gets all guides
      tags:
      - guide
    post:
      consumes:
      - application/json
      description: Creates a guide
      operationId: create-guide
      parameters:
      - description: Guide
        in: body
        name: guide
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guide'
        "400":
          description: Failed to create guide
          schema:
            type: string
      summary: Creates a guide
      tags:
      - guide
  /api/guides/{g_name}:
    delete:
      description: Deletes a guide
      operationId: delete-guide
      parameters:
      - description: GuideName
        in: path
        name: g_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Guide successfully deleted
          schema:
            type: string
        "404":
          description: Failed to delete guide
          schema:
            type: string
      summary: Deletes a guide
      tags:
      - guide
    get:
      description: Returns a guide from guide id
      operationId: get-guide
      parameters:
      - description: GuideName
        in: path
        name: g_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guide'
        "404":
          description: Failed to fetch guide
          schema:
            type: string
      summary: Gets a guide from guide id
      tags:
      - guide
    patch:
      consumes:
      - application/json
      description: Updates a guide
      operationId: update-guide
      parameters:
      - description: GuideName
        in: path
        name: g_name
        required: true
        type: string
      - description: Guide
        in: body
        name: guide
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guide'
        "400":
          description: Failed to update guide
          schema:
            type: string
      summary: Updates a guide
      tags:
      - guide
...
swagger: "2.0"